Split common conda env logic out of PythonRunner
